+++ lib/dado.dart
@@ -59,4 +59,5 @@ library dado;
 
 export 'src/binding.dart';
 export 'src/injector.dart';
+export 'src/key.dart';
 export 'src/module.dart';
