+++ lib/src/injector.dart
@@ -5,36 +5,10 @@
 library dado.injector;
 
 import 'dart:mirrors';
-import 'package:inject/inject.dart';
 import 'binding.dart';
+import 'key.dart';
+import 'module.dart';
 import 'utils.dart' as Utils;
-
-/**
- * Keys are used to resolve instances in an [Injector], they are used to
- * register bindings and request an object at the injection point.
- *
- * Keys consist of a [Symbol] representing the type name and an optional
- * annotation. If you need to create a Key from a [Type] literal, use [forType].
- */
-class Key {
-  final Symbol name;
-  final Object annotation;
-
-  Key(this.name, {Object annotatedWith}) : annotation = annotatedWith {
-    if (name == null) throw new ArgumentError("name must not be null");
-  }
-
-  factory Key.forType(Type type, {Object annotatedWith}) =>
-      new Key(Utils.typeName(type), annotatedWith: annotatedWith);
-
-  bool operator ==(o) => o is Key && o.name == name
-      && o.annotation == annotation;
-
-  int get hashCode => name.hashCode * 37 + annotation.hashCode;
-
-  String toString() => 'Key: $name'
-      '${(annotation!=null?' annotated with $annotation': '')}';
-}
 
 /**
  * An Injector constructs objects based on it's configuration. The Injector
@@ -228,59 +202,13 @@ class Injector {
 
   void _registerBindings(Type moduleType){
     var classMirror = reflectClass(moduleType);
-    var moduleMirror = classMirror.newInstance(const Symbol(''), []);
-
-    classMirror.declarations.values.forEach((member) {
-      if (member is VariableMirror) {
-        // Variables define "to instance" bindings
-        var instance = moduleMirror.getField(member.simpleName).reflectee;
-        var name = member.type.qualifiedName;
-        var annotation = Utils.getBindingAnnotation(member);
-        var key = new Key(name, annotatedWith: annotation);
-        _bindings[key] = new InstanceBinding(key, instance, moduleMirror);
-
-      } else if (member is MethodMirror) {
-        var name = member.returnType.qualifiedName;
-        var annotation = Utils.getBindingAnnotation(member);
-        Key key = new Key(name, annotatedWith: annotation);
-        if (member.isAbstract) {
-          if (member.isGetter) {
-            // Abstract getters define singleton bindings
-            _bindings[key] = new ConstructorBinding(key,
-                _selectConstructor(member.returnType),  moduleMirror,
-                singleton: true);
-          } else {
-            // Abstract methods define unscoped bindings
-            _bindings[key] = new ConstructorBinding(key,
-                _selectConstructor(member.returnType),  moduleMirror);
-          }
-        } else {
-          // Non-abstract methods produce instances by being invoked.
-          //
-          // In order for the method to use the injector to resolve dependencies
-          // it must be aware of the injector and the type we're trying to
-          // construct so we set the module's _currentInjector and
-          // _currentTypeName in the provider function.
-          //
-          // This is a slightly unfortunately coupling of Module to it's
-          // injector, but the only way we could find to make this work. It's
-          // a worthwhile tradeoff for having declarative bindings.
-          if (member.isGetter) {
-            // getters should define singleton bindings
-            _bindings[key] =
-                new ProviderBinding(key, member, moduleMirror,
-                    singleton: true);
-          } else {
-            // methods should define unscoped bindings
-            // TODO(justin): allow parameters in module method? This would make
-            // defining provided bindings much shorter when they rebind to a
-            // new type.
-            _bindings[key] =
-                new ProviderBinding(key, member, moduleMirror);
-          }
-        }
-      }
-    });
+    var module = classMirror.newInstance(const Symbol(''), []).reflectee;
+    
+    if (module is Module) {
+      _bindings.addAll(module.bindings);
+    } else {
+      throw new ArgumentError('$moduleType is not a Module');
+    }
   }
 
   void _verifyCircularDependency(Binding binding,
@@ -315,35 +243,6 @@ class Injector {
     });
 
     dependencyStack.removeLast();
-  }
-
-  MethodMirror _selectConstructor(ClassMirror m) {
-    Iterable<MethodMirror> constructors = Utils.getConstructorsMirrors(m);
-    // Choose contructor using @inject
-    MethodMirror ctor = constructors.firstWhere(
-      (c) => c.metadata.any(
-        (m) => m.reflectee == inject)
-      , orElse: () => null);
-
-    // In case there is no constructor annotated with @inject, see if there's a
-    // single constructor or a no-args.
-    if (ctor == null) {
-      if (constructors.length == 1) {
-        ctor = constructors.first;
-      } else {
-        ctor = constructors.firstWhere(
-            (c) => c.parameters.where((p) => !p.isOptional).length == 0
-        , orElse: () =>  null);
-      }
-    }
-
-    if (ctor == null) {
-      throw new ArgumentError("${m.qualifiedName} must have only "
-        "one constructor, a constructor annotated with @inject or no-args "
-        "constructor");
-    }
-
-    return ctor;
   }
 
   String toString() => 'Injector: $name';
